requirements.txt
requests
python-dotenv
openai
streamlit
streamlit-extras
plotly
pandas
openai
